[WIP] Separate CollectionCache
