+++ network-store-iidm-impl/src/main/java/com/powsybl/network/store/iidm/impl/CacheHandler.java
@@ -1,0 +1,5 @@
+package com.powsybl.network.store.iidm.impl;
+
+interface CacheHandler {
+    void init();
+}
